% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{\%>>>\%}
\alias{\%>>>\%}
\title{Play pipe}
\usage{
lhs \%>>>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics. the output order of the delayed expression can be identified using the .x# notation.}
}
\description{
Pipe a function - or call expression and runs the priorly paused evaluation
}
\details{
The pause pipe works like \code{\link[magrittr]{\%>\%}}, except the
return value is `rhs` as the second argument of the priorly paused evaluation
#'
}
\seealso{
\code{\link{\%//\%}}
}
